video-generation-agent/requirements.txt
# # Core Dependencies
# streamlit==1.45.0
# python-dotenv==1.0.0

# # Database and Caching
# redis==5.0.1
# pymongo==4.5.0

# # Monitoring & Logging
# prometheus-client==0.19.0
# structlog==24.1.0

# # HTTP & API
# requests==2.31.0
# aiohttp==3.9.1

# # Testing
# pytest==7.4.4
# pytest-asyncio==0.23.2

# # Utils
# python-dateutil==2.8.2
# spacy==3.8.0


# requirements.txt (수정 버전)
streamlit==1.45.0
python-dotenv==1.0.0
redis==5.0.1
pymongo==4.5.0
prometheus-client==0.19.0
structlog==24.1.0
requests==2.31.0
aiohttp==3.9.1
pytest==7.4.4
pytest-asyncio==0.23.2
python-dateutil==2.8.2
spacy==3.7.4  # 3.8.0 대신 안정 버전 사용
https://github.com/explosion/spacy-models/releases/download/ko_core_news_sm-3.7.0/ko_core_news_sm-3.7.0.tar.gz
numpy==1.26.4  # 명시적 버전 지정
pandas==2.0.3  # 경량 버전 사용
opencv-python-headless==4.9.0.80